+++ orchestrator/pkg/runner/terminator.go
@@ -3,5 +3,5 @@ package runner
 import "context"
 
 type Terminator interface {
-	Terminate(ctx context.Context, id int) error
+	Terminate(ctx context.Context, id uint64) error
 }
